Use EtaExpansion.liftApp for impure phantom arguments
