+++ compiler/src/dotty/tools/dotc/transform/VCInlineMethods.scala
@@ -44,7 +44,7 @@ class VCInlineMethods extends MiniPhaseTransform with IdentityDenotTransformer {
   override def phaseName: String = "vcInlineMethods"
 
   override def runsAfter: Set[Class[_ <: Phase]] =
-    Set(classOf[ExtensionMethods], classOf[PatternMatcher], classOf[PhantomArgumentEval])
+    Set(classOf[ExtensionMethods], classOf[PatternMatcher], classOf[PhantomArgLift])
 
   /** Replace a value class method call by a call to the corresponding extension method.
    *
